# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Stager API
  description: |
    API for Stager - AI-Powered Virtual Staging Platform for Real Estate.

    ## Authentication
    All endpoints (except webhooks) require authentication via Supabase Auth session cookies.
    Unauthenticated requests return `401 Unauthorized`.

    ## Rate Limiting
    Resource-intensive endpoints are rate-limited:
    - **Staging:** 20 requests/minute
    - **Email:** 10 requests/minute
    - **Auth:** 5 requests/minute

    Rate limit headers are included in responses:
    - `X-RateLimit-Limit`: Maximum requests allowed
    - `X-RateLimit-Remaining`: Requests remaining in window
    - `X-RateLimit-Reset`: Unix timestamp when limit resets

    ## Error Responses
    All errors follow a consistent format with actionable suggestions:
    ```json
    {
      "error": "Human-readable message",
      "code": "ERROR_CODE",
      "suggestion": "Actionable suggestion for the user",
      "action": { "label": "Button label", "href": "/path" },
      "details": { "field": "value" }
    }
    ```

    ## Credits
    - Staging costs 1 credit per image
    - First 2 remixes per original image are free
    - Additional remixes cost 1 credit each
  version: 1.136.0
  contact:
    name: Stager Support
servers:
  - url: https://stager-tau.vercel.app/api
    description: Production
  - url: http://localhost:3000/api
    description: Local Development
...
paths:
...
  /staging/versions:
    get:
      tags: [Staging]
      summary: Get all versions for a staging job
      description: Fetch all versions (remixes) linked to the same original image via version group.
      operationId: getStagingVersions
      parameters:
        - name: jobId
          in: query
          required: true
          schema:
            type: string
            format: uuid
          description: ID of any job in the version group
      responses:
        '200':
          description: Versions retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  versions:
                    type: array
                    items:
                      $ref: '#/components/schemas/StagingJobSummary'
                  versionGroupId:
                    type: string
                    format: uuid
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============ BILLING ============
...
  /billing/checkout:
    post:
      tags: [Billing]
      summary: Create checkout session
      description: Create a Stripe checkout session for subscribing to a plan.
      operationId: createCheckout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [planSlug]
              properties:
                planSlug:
                  type: string
                  enum: [standard, professional, enterprise]
                  description: Plan to subscribe to (cannot be 'free')
      responses:
        '200':
          description: Checkout session created
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                    format: uri
                    description: Stripe checkout URL to redirect user
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /properties/{id}/download:
    get:
      tags: [Properties]
      summary: Download property images as ZIP
      description: Download all staged images for a property as a ZIP file.
      operationId: downloadPropertyImages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Property ID
      responses:
        '200':
          description: ZIP file download
          content:
            application/zip:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============ ACCOUNT ============
  /account/delete:
    delete:
      tags: [Account]
      summary: Delete account
      description: |
        Permanently delete the user account and all associated data including:
        - Profile
        - Properties
        - Staging jobs
        - Stored images
        - Team memberships

        This action cannot be undone.
      operationId: deleteAccount
      responses:
        '200':
          description: Account deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ============ WEBHOOKS ============
...
components:
...
  schemas:
...
    RoomType:
      type: string
      enum:
        - living-room
        - bedroom-master
        - bedroom-guest
        - bedroom-kids
        - dining-room
        - kitchen
        - home-office
        - bathroom
        - outdoor-patio
      description: Type of room being staged
    FurnitureStyle:
      type: string
      enum:
        - modern
        - traditional
        - minimalist
        - mid-century
        - scandinavian
        - industrial
        - coastal
        - farmhouse
        - luxury
      description: Furniture style for staging
...
    JobStatus:
      type: string
      enum:
        - pending
        - queued
        - preprocessing
        - processing
        - uploading
        - completed
        - failed
      description: Current status of a staging job
    ErrorCode:
      type: string
      enum:
        - UNAUTHORIZED
        - SESSION_EXPIRED
        - FORBIDDEN
        - NOT_OWNER
        - INSUFFICIENT_PERMISSIONS
        - INSUFFICIENT_CREDITS
        - PAYMENT_REQUIRED
        - RATE_LIMITED
        - VALIDATION_ERROR
        - INVALID_INPUT
        - NOT_FOUND
        - ALREADY_EXISTS
        - CONFLICT
        - PROVIDER_ERROR
        - EMAIL_ERROR
        - INTERNAL_ERROR
        - SERVICE_UNAVAILABLE
      description: Standardized error codes

    # ---- Request Schemas ----
...
    ErrorResponse:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
          description: Human-readable error message
        code:
          $ref: '#/components/schemas/ErrorCode'
        suggestion:
          type: string
          description: Actionable suggestion for the user
        action:
          type: object
          description: Optional CTA button
          properties:
            label:
              type: string
            href:
              type: string
        details:
          type: object
          additionalProperties: true
          description: Additional context about the error
...
    StagingJobSummary:
      type: object
      description: Abbreviated staging job for lists and search results
      properties:
        id:
          type: string
          format: uuid
        roomType:
          $ref: '#/components/schemas/RoomType'
        style:
          $ref: '#/components/schemas/FurnitureStyle'
        stagedImageUrl:
          type: string
          format: uri
        status:
          $ref: '#/components/schemas/JobStatus'
        createdAt:
          type: string
          format: date-time
        property:
          type: object
          nullable: true
          properties:
            id:
              type: string
              format: uuid
            address:
              type: string
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "You must be logged in to perform this action."
            code: "UNAUTHORIZED"
            suggestion: "Please log in or create an account to continue."
            action:
              label: "Log In"
              href: "/login"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "The requested resource could not be found."
            code: "NOT_FOUND"
    ValidationError:
      description: Invalid request data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Please fix the errors in your submission."
            code: "VALIDATION_ERROR"
            suggestion: "Review the highlighted fields and correct any issues."
            details:
              fields:
                - field: "email"
                  message: "Invalid email format"
...